+++ RELEASENOTES.md
@@ -2,7 +2,9 @@
 
 ### 3.8.0 (UNRELEASED)
 * Dark mode.
+* Admin UI for users. [#340]
 * State/province lookup by menu. [#422]
+* Added check for semanticAdmin in the upgrade advisor. [#431]
 
 ### 3.7.4 (May 20, 2020)
 * Support for HY format.